temp

+++ thumbor/src/engine/photon.rs
@@ -197,7 +197,7 @@ impl SpecTransform<&Resize> for Photon {
 // }
 impl SpecTransform<&Watermark> for Photon {
     fn transform(&mut self, op: &Watermark) {
-        multiple::watermark(&mut self.0, &WATERMARK, op.x, op.y);
+        multiple::watermark(&mut self.0, &WATERMARK, op.x as i64, op.y as i64);
     }
 }
